# modules/module-billing-isp-api/openapi/v1/billing-isp.yaml
openapi: 3.1.0
info: {title: Billing ISP API, version: 1.0.0, description: Team-scoped access services, coverage, installation, RADIUS, usage, equipment, and network adapters.}
servers: [{url: /}]
paths:
  /api/v1/billing/isp:
    get:
      operationId: billing.isp.accessServices.list
      security: [{sanctum: []}]
      parameters: [{$ref: '#/components/parameters/PageSize'}]
      responses: {'200': {$ref: '#/components/responses/Collection'}, '401': {$ref: '#/components/responses/Unauthorized'}}
  /api/v1/billing/isp/capabilities:
    get: {operationId: billing.isp.capabilities.list, security: [{sanctum: []}], responses: {'200': {$ref: '#/components/responses/Collection'}}}
    post:
      operationId: billing.isp.capabilities.create
      security: [{sanctum: []}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CapabilityRequest'}}}}
      responses: {'201': {$ref: '#/components/responses/Document'}, '422': {$ref: '#/components/responses/ValidationError'}}
components:
  parameters:
    PageSize: {name: per_page, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
    IdempotencyKey: {name: Idempotency-Key, in: header, schema: {type: string, maxLength: 255}}
  responses:
    Collection: {description: Paginated team-scoped collection, content: {application/json: {schema: {type: object, required: [data, meta], properties: {data: {type: array, items: {type: object}}, meta: {type: object}}}}}}
    Document: {description: ISP resource, content: {application/json: {schema: {type: object, required: [data], properties: {data: {type: object}}}}}}
    Unauthorized: {description: Authentication or scope is required.}
    ValidationError: {description: RFC 9457-compatible validation problem.}
  schemas: {CapabilityRequest: {type: object, required: [capability], properties: {capability: {type: string}, provider: {type: string}, configuration: {type: object, additionalProperties: true}}}}
  securitySchemes: {sanctum: {type: http, scheme: bearer, bearerFormat: Sanctum}}